% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{summarize_ncopies}
\alias{summarize_ncopies}
\alias{pivot_wider_ncopies}
\title{Summarize community data by species, genus, family, etc.}
\usage{
summarize_ncopies(community_df, .by = .data$species, rm_unidentified = TRUE)

pivot_wider_ncopies(community_df, .by = .data$species, rm_unidentified = TRUE)
}
\arguments{
\item{community_df}{A community data frame,
e.g., output of \code{ranemone::read_tsv_xz("community_qc3nn_target.tsv.xz")}.}

\item{.by}{Columns by which to summarize, e.g., \code{species}, \code{genus}, and/or \code{family}.}

\item{rm_unidentified}{A logical value indicating whether to remove "unidentified" species.}
}
\value{
\code{summarize_ncopies()} returns a data frame in longer format.

\code{pivot_wider_ncopies()} returns a community composition matrix
in wider format (samples x species).
}
\description{
\code{ncopiesperml} is summed for in each sample.
Species whose names contain "unidentified" are removed by default.
}
\examples{
\dontrun{
community_df = ranemone::read_tsv_xz("community_qc3nn_target.tsv.xz")
species_df = ranemone::summarize_ncopies(community_df, .by = species)
genus_df = ranemone::summarize_ncopies(community_df, .by = genus, rm_unidentified = FALSE)

species_mat = ranemone::pivot_wider_ncopies(community_df, .by = species)
family_mat = ranemone::pivot_wider_ncopies(community_df, .by = family, rm_unidentified = FALSE)
}
}
