% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_dir}
\alias{cache_dir}
\title{Utilities for caching}
\usage{
cache_dir()
}
\value{
\code{cache_dir()} returns the path to the cache directory.
}
\description{
Some expensive operations are cached to avoid recomputing them.
Intermediate results are stored in a cache directory.
Users can set it via \code{ranemone.cache_dir} option.
The default is set with \code{\link[tools:R_user_dir]{tools::R_user_dir()}}, which persists across R sessions.
}
\details{
General recommendation is to set it to a persistent directory with easy access,
such as \verb{~/.cache/ranemone}, not a temporary directory like \code{\link[=tempdir]{tempdir()}}.
Setting it to \code{NA} or \code{FALSE} will make a new temporary directory each time
\code{\link[=cache_dir]{cache_dir()}} is called, which virtually disables caching (not recommended).
You can share the cache directory with other users on the same machine by
setting it to a common directory, such as \verb{/tmp/ranemone-cache}.
}
\examples{
old = options(ranemone.cache_dir = "~/.cache/ranemone-example")

ranemone::cache_dir()

\dontrun{
# Calculate the size
fs::dir_info(ranemone::cache_dir(), recurse = TRUE)$size |> sum()
}

options(old) # reset for this example, not needed in real use
}
