% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{wget_recursive}
\alias{wget_recursive}
\title{Download data from ANEMONE DB}
\usage{
wget_recursive(
  rel_url = "dist/",
  opts = c("--reject=.fastq.xz,.png", "--no-verbose"),
  ...,
  user = getOption("ranemone.user"),
  password = getOption("ranemone.password")
)
}
\arguments{
\item{rel_url}{A relative URL from the server root. If a directory is given,
it should end with a slash \code{/} for \code{--no-parent} to work as expected.}

\item{opts}{Additional options passed to \code{wget}.
\code{--recursive --level=0 --no-parent --continue} are preset.
\code{--directory-prefix} is set from \code{\link[=directory_prefix]{directory_prefix()}}.
\code{--http-user} and \code{--http-password} are set from \code{user} and \code{password} arguments,
respectively.
Currently \code{-N/--timestamping} does not work with the server:
"Last-modified header missing -- time-stamps turned off."}

\item{...}{Additional arguments passed to \code{\link[base:system2]{base::system2()}}.}

\item{user}{Your ANEMONE DB user name.
Setting it via \code{options(ranemone.user = "_")} will be easier for repeated use.}

\item{password}{"Data Download Password" generated in the
"Member Menu > Settings > Data Download" page, not your login password.
Setting it via \code{options(ranemone.password = "_")} will be easier for repeated use.}
}
\value{
The return value of \code{\link[base:system2]{base::system2()}}.
}
\description{
\code{wget} is used to download data files recursively from ANEMONE DB.
Your user name and a temporary password generated in the website are required.
}
\details{
The directory tree starting from \verb{db.anemone.bio/} will be created under
\code{\link[=directory_prefix]{directory_prefix()}} according to the default behavior of \code{wget -r},
which is assumed by other functions in this package.
}
\examples{
\dontrun{
options(ranemone.directory_prefix = "prefix/path/to/data")
options(ranemone.user = "your-username")
options(ranemone.password = "generated-password") # not yours

ranemone::wget_recursive("dist/MiFish/ANEMONE/2023KibanS/2023KibanSRUN01/")

fs::dir_ls(directory_prefix()) # "db.anemone.bio"
}
}
\seealso{
\url{https://db.anemone.bio/how-to-download-data/},
\url{https://www.gnu.org/software/wget/manual/wget.html}
}
