% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishbase.R
\name{read_fb_tbl}
\alias{read_fb_tbl}
\alias{fb_species_ecology_estimate}
\title{Utilities for FishBase data fetching and caching}
\usage{
read_fb_tbl(tbl, ..., force = FALSE)

fb_species_ecology_estimate(..., force = FALSE)
}
\arguments{
\item{tbl, ...}{Passed to \code{rfishbase::fb_tbl()}.}

\item{force}{Set to \code{TRUE} to ignore and overwrite existing cached files in \code{\link[=cache_dir]{cache_dir()}}.}
}
\value{
A data frame.

\code{fb_species_ecology_estimate()} selects essential columns from
"species", "ecology", and "estimate" tables.
}
\description{
Call \code{\link[rfishbase:fb_tbl]{rfishbase::fb_tbl()}} to fetch data from \href{https://www.fishbase.se/}{FishBase}
and cache the result in \code{\link[=cache_dir]{cache_dir()}}.
}
\details{
Note that some species have duplicated rows in ecology table,
which is resolved in \code{fb_species_ecology_estimate()}.
}
\examples{
\dontrun{
fb_species_ecology_estimate()

fb_species = read_fb_tbl("species")
fb_ecology = read_fb_tbl("ecology")
fb_estimate = read_fb_tbl("estimate")
}
}
