% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{data_path}
\alias{data_path}
\title{Get the path to the ANEMONE DB data directory.}
\usage{
data_path()
}
\value{
A path read from the option.
}
\description{
Set \code{ranemone.data_path} option to specify the path to the directory
containing ANEMONE DB data files, and use \code{data_path()} to get it.
It is used internally by other functions to locate the data files.
}
\examples{
old = options(ranemone.data_path = "path/to/your/data")

ranemone::data_path()

options(old) # reset for this example, not needed in real use
}
