% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranemone-package.R
\docType{package}
\name{ranemone-package}
\alias{ranemone}
\alias{ranemone-package}
\title{ranemone: R Interface to ANEMONE DB}
\description{
ANEMONE DB is a database for biodiversity surveys based on environmental DNA metabarcoding. It is managed by Dr. Michio Kondoh, Ecological Integration Lab. at Tohoku University. This package provides an utility to analyze data from ANEMONE DB.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/heavywatal/ranemone}
  \item \url{https://heavywatal.github.io/ranemone/}
  \item Report bugs at \url{https://github.com/heavywatal/ranemone/issues}
}

}
\author{
\strong{Maintainer}: Watal M. Iwasaki \email{heavywatal@gmail.com} (\href{https://orcid.org/0000-0002-4718-7605}{ORCID})

Authors:
\itemize{
  \item Watal M. Iwasaki \email{heavywatal@gmail.com} (\href{https://orcid.org/0000-0002-4718-7605}{ORCID})
}

}
\keyword{internal}
