% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{summarize_alpha_diversity}
\alias{summarize_alpha_diversity}
\alias{calc_alpha_div}
\title{Calculate alpha diversity indices}
\usage{
summarize_alpha_diversity(species_df, .by = .data$samplename, .groups = NULL)

calc_alpha_div(mat)
}
\arguments{
\item{species_df}{A data frame summarized by species,
e.g., output of \code{\link[=summarize_ncopies]{summarize_ncopies()}}.}

\item{.by}{Passed to \code{\link[dplyr:summarise]{dplyr::summarize()}}.}

\item{.groups}{Passed to \code{\link[dplyr:summarise]{dplyr::summarize()}}.}

\item{mat}{A community composition matrix (samples x species),
e.g., output of \code{\link[=pivot_wider_ncopies]{pivot_wider_ncopies()}}.}
}
\value{
A data frame with alpha diversity indices.
}
\description{
Alpha diversity indices are calculated using \code{ncopiesperml} as a proxy of abundances.
}
\details{
The output from these two functions should be equivalent given the same input.
\code{summarize_alpha_diversity()} operates in a tidyverse style with longer data frames,
while \code{calc_alpha_div()} operates on a community composition matrix in wider format
using the \href{https://cran.r-project.org/package=vegan}{vegan} package.
}
\examples{
\dontrun{
species_df = ranemone::summarize_ncopies(community_df, .by = species)
res1 = ranemone::summarize_alpha_diversity(species_df)

species_mat = ranemone::pivot_wider_ncopies(community_df, .by = species)
res2 = ranemone::calc_alpha_div(species_mat)

stopifnot(all.equal(res1, res2))
}
}
