% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{directory_prefix}
\alias{directory_prefix}
\title{Get the prefix path to the ANEMONE DB data directory.}
\usage{
directory_prefix()
}
\value{
A path read from the option.
}
\description{
\code{directory_prefix()} reads the \code{ranemone.directory_prefix} option set by users,
and is used for downloading and reading files.
It should be a persistent directory with easy access, such as \verb{~/db},
not a temporary directory like \code{\link[=tempdir]{tempdir()}}.
}
\details{
If not set, it falls back to the \code{extdata/directory_prefix} in the package
installation directory, which is mainly for testing and examples.
}
\examples{
old = options(ranemone.directory_prefix = "prefix/path/to/data")

ranemone::directory_prefix()

options(old) # reset for this example, not needed in real use
}
