% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_tsv_xz}
\alias{read_tsv_xz}
\title{Read files recursively in the ANEMONE DB data directory.}
\usage{
read_tsv_xz(filename, ..., limit = 250L)
}
\arguments{
\item{filename}{A filename to read.
Files are searched recursively in the path returned by \code{\link[=data_path]{data_path()}}.
It should be one of the following:
\itemize{
\item \code{community_qc3nn_target.tsv.xz}
\item \code{community_qc_target.tsv.xz}
\item \code{community_standard.tsv.xz}
\item \code{experiment.tsv.xz}
\item \code{sample.tsv.xz}
}}

\item{...}{Additional arguments passed to \code{\link[readr:read_delim]{readr::read_tsv()}}.}

\item{limit}{Maximum number of files to read at once.}
}
\value{
A concatenated data frame.
}
\description{
ANEMONE DB contains a large collection of small tables,
and this package provides efficient ways to read and process them.
It may take a while to run them for the first time,
but subsequent runs will be very quick because the results are cached.
}
\examples{
\dontrun{

ranemone::read_tsv_xz("sample.tsv.xz")
ranemone::read_tsv_xz("experiment.tsv.xz")
ranemone::read_tsv_xz("community_qc3nn_target.tsv.xz")

}
}
